refactor: move ExecutorContextMetadata to distri-types, use in CLI directly
