+++ distri-cli/src/message.rs
@@ -2,8 +2,8 @@ use distri::Distri;
 use distri_a2a::{
     EventKind, Message as A2aMessage, MessageSendParams, Part as A2aPart, Role, TextPart,
 };
-use distri_types::RuntimeMode;
-use serde::Serialize;
+use distri_types::configuration::DefinitionOverrides;
+use distri_types::{ExecutorContextMetadata, RuntimeMode};
 use std::collections::HashMap;
 
 /// Build a lightweight connections summary to inject into the agent's prompt context.
@@ -36,35 +36,6 @@ pub async fn build_connections_context(client: &Distri) -> Option<String> {
     Some(lines.join("\n"))
 }
 
-/// Typed metadata for CLI requests. Serializes to the same JSON shape
-/// that the server's `ExecutorContextMetadata` expects.
-#[derive(Debug, Serialize, Default)]
-struct CliRequestMetadata {
-    /// Runtime mode — always "cli" for the CLI client.
-    runtime_mode: RuntimeMode,
-
-    /// Optional definition overrides (e.g. model override).
-    #[serde(skip_serializing_if = "Option::is_none")]
-    definition_overrides: Option<DefinitionOverridesPartial>,
-
-    /// Dynamic key-value pairs available in prompt templates.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    dynamic_values: Option<HashMap<String, serde_json::Value>>,
-
-    /// Environment variables passed to the server for execution.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    env_vars: Option<HashMap<String, String>>,
-}
-
-/// Subset of definition overrides the CLI can set.
-#[derive(Debug, Serialize)]
-struct DefinitionOverridesPartial {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    model: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    remote: Option<bool>,
-}
-
 pub fn build_message_params(
     content: String,
     thread_id: Option<&str>,
@@ -85,18 +56,19 @@ pub fn build_message_params_full(
     env_vars: Option<HashMap<String, String>>,
 ) -> MessageSendParams {
     let has_overrides = model.is_some() || remote;
-    let mut metadata = CliRequestMetadata {
+    let mut metadata = ExecutorContextMetadata {
         runtime_mode: RuntimeMode::Cli,
         definition_overrides: if has_overrides {
-            Some(DefinitionOverridesPartial {
+            Some(DefinitionOverrides {
                 model: model.map(|m| m.to_string()),
                 remote: if remote { Some(true) } else { None },
+                ..Default::default()
             })
         } else {
             None
         },
-        dynamic_values: None,
         env_vars,
+        ..Default::default()
     };
 
     if let Some(conn_ctx) = connections_context {
