+++ server/distri-core/src/agent/context.rs
@@ -2,11 +2,10 @@ use crate::agent::pricing;
 use crate::hooks_runtime::HookRegistry;
 use distri_stores::SessionStoreExt;
 use distri_types::{
-    configuration::DefinitionOverrides, AgentContextSize, AgentPlan, ContextBudget, ContextSize,
-    ContextUsage, ExecutionHistoryEntry, ExecutionResult, ModelSettings, Part, PlanStep, RunUsage,
+    AgentContextSize, AgentPlan, ContextBudget, ContextSize, ContextUsage,
+    ExecutionHistoryEntry, ExecutionResult, ModelSettings, Part, PlanStep, RunUsage,
     ScratchpadEntry, ScratchpadEntryType,
 };
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use std::{
     collections::{HashMap, HashSet},
@@ -18,7 +17,7 @@ use crate::agent::prompt_registry::PromptSection;
 use crate::{
     servers::registry::McpServerRegistry,
     tools::Tool,
-    types::{ExternalTool, Message},
+    types::Message,
     AgentError, AgentOrchestrator,
 };
 
@@ -42,55 +41,7 @@ pub enum ForkType {
     NewThread,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ExecutorContextMetadata {
-    /// Add additional context for tools to use passed as meta in tool calls
-    pub tool_metadata: Option<std::collections::HashMap<String, serde_json::Value>>,
-    #[serde(default)]
-    pub metadata: Option<serde_json::Value>,
-    #[serde(default)]
-    pub additional_attributes: Option<AdditionalAttributes>,
-
-    /// Define additional tools to delegate to during execution
-    pub external_tools: Option<Vec<ExternalTool>>,
-
-    /// Optional definition overrides supplied by the client
-    #[serde(default)]
-    pub definition_overrides: Option<DefinitionOverrides>,
-
-    /// Dynamic prompt sections to inject into the template per-call
-    #[serde(default)]
-    pub dynamic_sections: Option<Vec<PromptSection>>,
-
-    /// Dynamic key-value pairs available in templates per-call
-    #[serde(default)]
-    pub dynamic_values: Option<HashMap<String, serde_json::Value>>,
-
-    /// Browser session ID for browser tool integration
-    #[serde(default)]
-    pub browser_session_id: Option<String>,
-
-    /// Environment variables passed from the client to be available during execution.
-    /// These are forwarded to skill scripts and plugin contexts alongside secrets.
-    #[serde(default)]
-    pub env_vars: Option<HashMap<String, String>>,
-
-    /// When true, unsafe tools are simulated via LLM instead of executed.
-    /// Set by the eval simulator to test agent behavior without side effects.
-    #[serde(default)]
-    pub dry_run: Option<bool>,
-
-    /// Runtime environment. Determines which built-in agent variants to use.
-    /// Set by the client: "cli" for distri-cli, "browser" for JS SDK, defaults to "cloud".
-    #[serde(default)]
-    pub runtime_mode: distri_types::RuntimeMode,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct AdditionalAttributes {
-    pub thread: Option<serde_json::Value>,
-    pub task: Option<serde_json::Value>,
-}
+pub use distri_types::{AdditionalAttributes, ExecutorContextMetadata};
 
 #[derive(Debug, Clone, Default)]
 pub struct PromptTemplateOverride {
